% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{abs_error}
\alias{abs_error}
\title{Absolute Error}
\usage{
abs_error(pred, truth)
}
\arguments{
\item{pred}{a numeric vector of predicted values}

\item{truth}{a numeric vector of true values}
}
\value{
a numeric vector of absolute errors
}
\description{
Absolute Error
}
