% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{bin_chopper}
\alias{bin_chopper}
\title{Bin chopper function for binned bootstrapping}
\usage{
bin_chopper(x, nbins, return_breaks = FALSE)
}
\arguments{
\item{x}{vector of values to be binned}

\item{nbins}{number of bins}

\item{return_breaks}{logical indicating whether to return the bin breaks}
}
\description{
Bin chopper function for binned bootstrapping
}
