% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{bindividual_alpha}
\alias{bindividual_alpha}
\title{Bin-Individual Alpha Function for Conformal Prediction}
\usage{
bindividual_alpha(minqs, alpha)
}
\arguments{
\item{minqs}{Minimum quantiles}

\item{alpha}{alpha level}
}
\description{
Bin-Individual Alpha Function for Conformal Prediction
}
\keyword{internal}
