% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_predictions.R
\name{calibrate_predictions}
\alias{calibrate_predictions}
\title{Calibrate predictions using a calibration dataset}
\usage{
calibrate_predictions(
  pred,
  calib,
  calib_truth = NULL,
  method = c("glm", "isotonic"),
  family = "gaussian",
  transform = NULL
)
}
\arguments{
\item{pred}{Vector of predicted values}

\item{calib}{A numeric vector of predicted values in the calibration partition or a 2 or 3 column tibble or matrix with the first column being the predicted values and the second column being the truth values and the third, optional, column being the bin labels. If calib is a numeric vector, calib_truth and either calib_bins or breaks must be provided.}

\item{calib_truth}{A numeric vector of true values in the calibration partition}

\item{method}{Method for calibration, either 'glm' for generalized linear model or 'isotonic' for isotonic regression}

\item{family}{Family for the generalized linear model passed to `glm`. Default is 'gaussian' for a linear model. Other options include 'binomial', 'poisson', etc.}

\item{transform}{Optional transformation function, or string matching a function, to apply to the predicted values before calibration. If NULL, no transformation is applied. See details below.}
}
\value{
A numeric vector of calibrated predictions, aligned with the input vector `pred`. If method = 'glm' is used, the regression coefficients, family, and transformation from the fitted model can be accessed as attributes of the returned vector.
}
\description{
This function calibrates a vector of predicted values using a held-out calibration set of predictions and true outcomes. It supports both parametric calibration via generalized linear models (`glm`) and non-parametric calibration via isotonic regression. The function returns a vector of calibrated predictions.
}
\details{
This function calibrates model predictions to better match the distribution of observed outcomes, improving reliability especially when prediction error varies systematically. When `method = "glm"`, a generalized linear model is fit using the calibration set. When `method = "isotonic"`, a monotonic fit is estimated to map predicted values to expected outcomes. Platt scaling is a special case of calibration using `glm` with a logistic link function, which is useful for binary classification tasks.

If a transformation function is provided via `transform`, it is applied to both the prediction vector (`pred`) and the calibration predictions (`calib`) before calibration is performed. This is useful in certain cases, for exampel if the predictions are on a probability scale, it can be beneficial to apply a logit transformation before calibration.
}
\examples{
# Generate synthetic example
set.seed(123)
x <- runif(1000)
y <- rlnorm(1000, meanlog = 1 + x, sdlog = 0.5)
df <- tibble::tibble(x, y)

# Split into training, calibration, and test sets
df_train <- df[1:500, ]
df_cal <- df[501:750, ]
df_test <- df[751:1000, ]

# Fit a model and predict
model <- lm(log(y) ~ x, data = df_train)
pred_cal <- exp(predict(model, newdata = df_cal))
pred_test <- exp(predict(model, newdata = df_test))

# Calibrate using isotonic regression
calibrated_iso <- calibrate_predictions(
  pred = pred_test,
  calib = pred_cal,
  calib_truth = df_cal$y,
  method = "isotonic"
)

#' # Calibrate using GLM
calibrated_glm <- calibrate_predictions(
  pred = pred_test,
  calib = pred_cal,
  calib_truth = df_cal$y,
  method = "glm",
  family = "gaussian")

# Calibrate using GLM with a log transform
calibrated_glm <- calibrate_predictions(
  pred = pred_test,
  calib = pred_cal,
  calib_truth = df_cal$y,
  method = "glm",
  family = "gaussian",
  transform = log
)
}
