% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cauchy_kern}
\alias{cauchy_kern}
\title{Cauchy Kernel Function}
\usage{
cauchy_kern(d)
}
\arguments{
\item{d}{a numeric vector of distances}
}
\value{
a numeric vector of Cauchy kernel values
}
\description{
Cauchy Kernel Function
}
