% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{contiguize_intervals}
\alias{contiguize_intervals}
\title{Contiguize non-contiguous intervals}
\usage{
contiguize_intervals(
  pot_lower_bounds,
  pot_upper_bounds,
  empirical_lower_bounds,
  empirical_upper_bounds,
  return_all = FALSE
)
}
\arguments{
\item{pot_lower_bounds}{Potential non-contiguous lower bounds}

\item{pot_upper_bounds}{Potential non-contiguous upper bounds}

\item{empirical_lower_bounds}{Observed lower bounds}

\item{empirical_upper_bounds}{Observed upper bounds}

\item{return_all}{Return all intervals or just contiguous intervals}
}
\description{
Contiguize non-contiguous intervals
}
