% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{Dm_finder}
\alias{Dm_finder}
\title{Function to Find the Minimum Distance Between a Class and a Set of Clusters}
\usage{
Dm_finder(
  ncs,
  class_vec,
  class,
  clusters,
  return = c("min", "which.min", "vec")
)
}
\arguments{
\item{ncs}{Vector of non-conformity scores}

\item{class_vec}{Vector of class labels}

\item{class}{Class label to compare against the clusters}

\item{clusters}{List of clusters}

\item{return}{Character string indicating what to return. Options are 'min' for the minimum distance, 'which.min' for the index of the cluster with the minimum distance, or 'vec' for a vector of distances to each cluster.}
}
\value{
A numeric value or vector depending on the value of the `return` parameter. If `return` is 'min', returns the minimum distance. If `return` is 'which.min', returns the index of the cluster with the minimum distance. If `return` is 'vec', returns a vector of distances to each cluster.
}
\description{
Function to Find the Minimum Distance Between a Class and a Set of Clusters
}
\keyword{internal}
