% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{dm_to_prob}
\alias{dm_to_prob}
\title{Function to Convert Distance Measure to Probability}
\usage{
dm_to_prob(dm, dms)
}
\arguments{
\item{dm}{Distance measure}

\item{dms}{Vector of distance measures for all clusters}
}
\value{
A numeric value representing the probability of the distance measure relative to the sum of all distance measures
}
\description{
Function to Convert Distance Measure to Probability
}
\keyword{internal}
