% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gauss_kern}
\alias{gauss_kern}
\title{Gaussian Kernel Function}
\usage{
gauss_kern(d)
}
\arguments{
\item{d}{a numeric vector of distances}
}
\value{
a numeric vector of Gaussian kernel values
}
\description{
Gaussian Kernel Function
}
\keyword{internal}
