% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{heterogeneous_error}
\alias{heterogeneous_error}
\title{Heterogeneous Error Function for Non-Conformity Scores}
\usage{
heterogeneous_error(pred, truth, coefs)
}
\arguments{
\item{pred}{a numeric vector of predicted values}

\item{truth}{a numeric vector of true values}

\item{coefs}{a numeric vector of coefficients for the heterogeneous error model. Must be of length 2, where the first element is the intercept and the second element is the slope.}
}
\description{
Heterogeneous Error Function for Non-Conformity Scores
}
\keyword{internal}
