% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{kmeans_cluster_qecdf}
\alias{kmeans_cluster_qecdf}
\title{Function to Perform K-Means Clustering on Quantile Empirical Cumulative Distribution Functions (qECDFs) of Non-Conformity Scores}
\usage{
kmeans_cluster_qecdf(ncs, class_vec, q = seq(0.1, 0.9, by = 0.1), m)
}
\arguments{
\item{ncs}{Vector of non-conformity scores}

\item{class_vec}{Vector of class labels}

\item{q}{Quantiles to use for the qECDFs, default is a sequence from 0.1 to 0.9 in steps of 0.1}

\item{m}{Number of clusters to form}
}
\value{
A list of clusters, where each element is a vector of class labels assigned to that cluster
}
\description{
Function to Perform K-Means Clustering on Quantile Empirical Cumulative Distribution Functions (qECDFs) of Non-Conformity Scores
}
\keyword{internal}
