% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ks_cluster}
\alias{ks_cluster}
\title{Function to perform Kolmogorov-Smirnov clustering on non-conformity scores}
\usage{
ks_cluster(ncs, class_vec, m, maxit = 100, nrep = 10)
}
\arguments{
\item{ncs}{Vector of non-conformity scores}

\item{class_vec}{Vector of class labels}

\item{m}{Number of clusters to form}

\item{maxit}{Maximum number of iterations for the clustering algorithm}

\item{nrep}{Number of repetitions for the clustering algorithm}
}
\value{
A vector of cluster assignments, with attributes containing the clusters, coverage gaps, method used, number of clusters, and Calibrated Clustering index
}
\description{
Function to perform Kolmogorov-Smirnov clustering on non-conformity scores
}
