% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ks_cluster_assignment_step}
\alias{ks_cluster_assignment_step}
\title{Function to assign classes to clusters based on Kolmogorov-Smirnov clustering}
\usage{
ks_cluster_assignment_step(ncs, class_vec, class_labels, clusters, m)
}
\arguments{
\item{ncs}{Vector of non-conformity scores}

\item{class_vec}{Vector of class labels}

\item{class_labels}{Vector of unique class labels}

\item{clusters}{List of clusters}

\item{m}{Number of clusters}
}
\value{
A list of clusters, where each element is a vector of class labels assigned to that cluster
}
\description{
Function to assign classes to clusters based on Kolmogorov-Smirnov clustering
}
