% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{logistic_kern}
\alias{logistic_kern}
\title{Logistic Kernel Function}
\usage{
logistic_kern(d)
}
\arguments{
\item{d}{a numeric vector of distances}
}
\value{
a numeric vector of logistic kernel values
}
\description{
Logistic Kernel Function
}
