% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{minq_to_alpha}
\alias{minq_to_alpha}
\title{Helper for minimum quantile to alpha function}
\usage{
minq_to_alpha(minq, alpha)
}
\arguments{
\item{minq}{minimum quantile}

\item{alpha}{alpha level}
}
\description{
Helper for minimum quantile to alpha function
}
