% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pintervals-package.R
\docType{package}
\name{pintervals-package}
\alias{pintervals}
\alias{pintervals-package}
\title{pintervals: Model Agnostic Prediction Intervals}
\description{
Provides tools for estimating model-agnostic prediction intervals using conformal prediction, bootstrapping, and parametric prediction intervals. The package is designed for ease of use, offering intuitive functions for both binned and full conformal prediction methods, as well as parametric interval estimation with diagnostic checks. Currently only working for continuous predictions.
}
\author{
\strong{Maintainer}: David Randahl \email{david.randahl@pcr.uu.se}

}
\keyword{internal}
