% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{raw_error}
\alias{raw_error}
\title{Raw Error Function for Non-Conformity Scores}
\usage{
raw_error(pred, truth)
}
\arguments{
\item{pred}{a numeric vector of predicted values}

\item{truth}{a numeric vector of true values}
}
\value{
a numeric vector of raw errors
}
\description{
Raw Error Function for Non-Conformity Scores
}
