% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{reciprocal_linear_kern}
\alias{reciprocal_linear_kern}
\title{Reciprocal Linear Kernel Function}
\usage{
reciprocal_linear_kern(d)
}
\arguments{
\item{d}{a numeric vector of distances}
}
\value{
a numeric vector of reciprocal linear kernel values
}
\description{
Reciprocal Linear Kernel Function
}
