% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rel_error}
\alias{rel_error}
\title{Relative Error Function for Non-Conformity Scores by predicted Values}
\usage{
rel_error(pred, truth)
}
\arguments{
\item{pred}{a numeric vector of predicted values}

\item{truth}{a numeric vector of true values}
}
\value{
a numeric vector of relative errors
}
\description{
Relative Error Function for Non-Conformity Scores by predicted Values
}
