% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{resolve_weight_function}
\alias{resolve_weight_function}
\title{Resolve weight function from string or function}
\usage{
resolve_weight_function(weight_function)
}
\arguments{
\item{weight_function}{A character string or function}
}
\value{
A weight function
}
\description{
Resolve weight function from string or function
}
\keyword{internal}
