% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{wcss_compute}
\alias{wcss_compute}
\title{Function to Compute the Within-Cluster Sum of Squares (WCSS) for a Set of Clusters}
\usage{
wcss_compute(ncs, class_vec, cluster, q = seq(0.1, 0.9, by = 0.1))
}
\arguments{
\item{ncs}{Vector of non-conformity scores}

\item{class_vec}{Vector of class labels}

\item{cluster}{Vector of cluster labels}

\item{q}{Quantiles to use for the qECDFs, default is a sequence from 0.1 to 0.9 in steps of 0.1}
}
\value{
A numeric value representing the WCSS for the cluster
}
\description{
Function to Compute the Within-Cluster Sum of Squares (WCSS) for a Set of Clusters
}
\keyword{internal}
