% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{weights_calculator}
\alias{weights_calculator}
\title{Weights calculator for weighted conformal prediction}
\usage{
weights_calculator(y_hat, calib)
}
\arguments{
\item{y_hat}{Predicted value}

\item{calib}{a vector of true values of the calibration partition}
}
\description{
Weights calculator for weighted conformal prediction
}
