% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{bindividual_alpha}
\alias{bindividual_alpha}
\title{Bin-individual alpha function for conformal prediction}
\usage{
bindividual_alpha(minqs, alpha)
}
\arguments{
\item{minqs}{Minimum quantiles}

\item{alpha}{alpha level}
}
\description{
Bin-individual alpha function for conformal prediction
}
