% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{bootstrap_inner}
\alias{bootstrap_inner}
\title{Bootstrap function for bootstrapping the prediction intervals}
\usage{
bootstrap_inner(pred, error, nboot, alpha, lower_bound, upper_bound)
}
\arguments{
\item{pred}{predicted value}

\item{error}{vector of errors}

\item{nboot}{number of bootstrap samples}

\item{alpha}{confidence level}

\item{lower_bound}{lower bound of the prediction interval}

\item{upper_bound}{upper bound of the prediction interval}
}
\value{
a numeric vector with the predicted value and the lower and upper bounds of the prediction interval
}
\description{
Bootstrap function for bootstrapping the prediction intervals
}
