% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{class_to_clusters}
\alias{class_to_clusters}
\title{Function to Convert Class Vector to Cluster Vector Based on Calibrated Clusters}
\usage{
class_to_clusters(class_vec, cluster_vec_calib)
}
\arguments{
\item{class_vec}{Vector of class labels}

\item{cluster_vec_calib}{Vector of calibrated clusters}
}
\value{
A vector of cluster assignments, with attributes containing the clusters, method used, number of clusters, Calibrated Clustering index, and coverage gaps
}
\description{
Function to Convert Class Vector to Cluster Vector Based on Calibrated Clusters
}
\keyword{internal}
