% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{coverage_gap_finder}
\alias{coverage_gap_finder}
\title{Function to Find the Coverage Gap for a Set of Clusters}
\usage{
coverage_gap_finder(ncs, class_vec, cluster)
}
\arguments{
\item{ncs}{Vector of non-conformity scores}

\item{class_vec}{Vector of class labels}

\item{cluster}{Vector of cluster labels}
}
\value{
A numeric value representing the maximum coverage gap between the clusters
}
\description{
Function to Find the Coverage Gap for a Set of Clusters
}
\keyword{internal}
