% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{flatten_cp_bin_intervals}
\alias{flatten_cp_bin_intervals}
\title{Flatten binned conformal prediction intervals to contiguous intervals}
\usage{
flatten_cp_bin_intervals(lst, contiguize = FALSE)
}
\arguments{
\item{lst}{list of binned conformal prediction intervals}

\item{contiguize}{logical indicating whether to contiguize the intervals}
}
\description{
Flatten binned conformal prediction intervals to contiguous intervals
}
