% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ks_cluster_init_step}
\alias{ks_cluster_init_step}
\title{Function to initialize clusters for Kolmogorov-Smirnov clustering}
\usage{
ks_cluster_init_step(ncs, class_vec, m)
}
\arguments{
\item{ncs}{Vector of non-conformity scores}

\item{class_vec}{Vector of class labels}

\item{m}{Number of clusters to form}
}
\description{
Function to initialize clusters for Kolmogorov-Smirnov clustering
}
