% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformalprediction-package.R
\docType{package}
\name{pintervals-package}
\alias{pintervals}
\alias{pintervals-package}
\title{pintervals: Model agnostic prediction intervals}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: David Randahl \email{david.randahl@pcr.uu.se}

Authors:
\itemize{
  \item Jonathan Williams
}

Other contributors:
\itemize{
  \item Håvard Hegre [contributor]
}

}
\keyword{internal}
