% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{za_rel_error}
\alias{za_rel_error}
\title{Zero-adjusted Relative Error Function for Non-Conformity Scores by predicted Values with a small adjustment}
\usage{
za_rel_error(pred, truth)
}
\arguments{
\item{pred}{a numeric vector of predicted values}

\item{truth}{a numeric vector of true values}
}
\value{
a numeric vector of zero-adjusted relative errors
}
\description{
Zero-adjusted Relative Error Function for Non-Conformity Scores by predicted Values with a small adjustment
}
